requirements.txt
coverage
openai
GitPython
asyncio
requests
pytest
transformers
# torch
fastapi
pydantic
uvicorn[standard]
huggingface_hub
# torch==2.5.0
# torchvision==0.20.0
# torchaudio==2.5.0
# -f https://download.pytorch.org/whl/torch_stable.html
boto3>=1.28.0
aioboto3>=12.0.0
python-dotenv>=1.0.0
freezegun
pytest-asyncio